Consolidate source files and adopt ParadiseECS editorconfig
Merge ITypeInfo.cs and TypeInfoRegistry.cs into OneShot.cs to keep the
runtime library as a single source file. Update XML doc comments with
<see cref> cross-references and clarify the fluent builder pipeline
(WithBuilder -> LifetimeBuilder -> ResolverBuilder). Add missing XML
docs for InjectAttribute, ILabel<T>, LabelExtension, and MethodInfoCache.

Adopt the ParadiseECS .editorconfig which adds public field naming rules,
additional CA diagnostics, and stricter readonly field enforcement.

Target netstandard2.1 for broader compatibility and replace
ArgumentNullException.ThrowIfNull with manual null checks.
